加载模块：基本加载使用提供者SPI
