+++ data-provider/data-provider-base/src/main/java/com/cn/jmw/DataSourceManager.java
@@ -1,0 +1,10 @@
+package com.cn.jmw;
+
+/**
+ * @author jmw
+ * @Description TODO
+ * @date 2023年04月09日 18:22
+ * @Version 1.0
+ */
+public class DataSourceManager {
+}
